Finished writing code for displaying etds
Finished writing code to render the etds; hopefully now with less bugs.
There is enough code for a mvp, but it will suck.

Basic features missing:
-Error checking for user’s input in the station input activity
-Swiping up to refresh estimates
-Ability for user to edit their data

Notes:
-I will most likely not being using animations past the onboarding
stage. If I were using this app and the majority of the time spent
waiting for simple data was used to wait for animations to finish, I’d
be pretty ticked off.