Notification code completed (for now)
- Dialog will now pop up to prompt user how early they want to set the
alarm for
- Dialog will also detect if user wants to set their alarm too early
and will not start an alarm
- Fixed bug (to the best of my ability) related to the timer in the
notification “ending early”. (It’s not ending early, it’s ending on
time. It looks like it’s ending early because ticks are often times a
few ms more than specified which gives the illusion of the timer ending
early. Semi-fixed this by starting another timer in the current one if
it “ends early”).
- Renamed EtdsFragment to DeparturesFragment