Refactored MyStationsFragment
- Refactored MyStationsFragment based on feedback
- Saving changes is now done through lazy loading. User may assume that
all changes are saved on the spot

Due to the complexity of the BartDataElem view, it’s incredibly complex
to implement MyStationsFragment as a RecyclerView, so I’m the design of
appending views in a ScrollView alone.

Unable to utilize RecyclerViews for this project since a lot of the
adapters are being used for non-RecyclerView things.