Fixed bug related to crashing on update
- Fixed bug relating to the app crashing whenever the user’s route is
updated. This bug is caused from failing to update the local copy of
the user’s persisted data, leading to a bad filtering to be preformed
whenever new real time estimates are fetched (everything is filtered
out). Because of this, whenever we attempt to retrieve the first Etd
for each EtdStaiton, it crashes (as 0 is out of bounds for an empty
list).